Fail closed on WES when require_auth is enabled.
Anonymous list/submit and related run routes now return 401 under pilot auth, and the pilot-local issuer matches host-issued broker tokens so Bearer Passports validate.

+++ crates/ferrum-core/src/disk.rs
@@ -42,7 +42,7 @@ fn platform_free_space(path: &Path) -> Option<(u64, u64)> {
     if rc != 0 {
         return None;
     }
-    let block_size = stat.f_frsize as u64;
+    let block_size = stat.f_frsize;
     let total = stat.f_blocks as u64 * block_size;
     let free = stat.f_bavail as u64 * block_size;
     Some((total, free))
